+++ projetoSqueezy/package.json
@@ -20,7 +20,8 @@
     "@fontsource/poppins": "^5.0.14",
     "@react-native-async-storage/async-storage": "^1.23.1",
     "@react-native-picker/picker": "^2.7.7",
-    "@react-navigation/native": "^6.0.2",
+    "@react-navigation/native": "^6.1.17",
+    "@react-navigation/native-stack": "^6.9.26",
     "expo": "~51.0.11",
     "expo-app-loading": "^2.1.1",
     "expo-constants": "~16.0.2",
@@ -40,8 +41,8 @@
     "react-native": "^0.74.2",
     "react-native-gesture-handler": "~2.16.1",
     "react-native-reanimated": "~3.10.1",
-    "react-native-safe-area-context": "4.10.1",
-    "react-native-screens": "3.31.1",
+    "react-native-safe-area-context": "^4.10.1",
+    "react-native-screens": "^3.31.1",
     "react-native-uuid": "^2.0.2",
     "react-native-web": "~0.19.10"
   },
